+++ src/main/java/com/comp354pjb/codenames/model/Observer.java
@@ -1,4 +1,4 @@
-package com.comp354pjb.codenames.Model;
+package com.comp354pjb.codenames.model;
 
 public interface Observer
 {
