+++ src/main/java/com/comp354pjb/codenames/model/player/OperativeAI.java
